eews_backend/seedlink/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
confluent-kafka==2.2.0
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
fonttools==4.42.1
greenlet==2.0.2
idna==3.4
influxdb-client==1.37.0
kiwisolver==1.4.4
lxml==4.9.3
matplotlib==3.7.2
numpy==1.25.2
obspy==1.4.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
reactivex==4.0.4
requests==2.31.0
scipy==1.11.2
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
